% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tp_get_info.R
\name{tp_get_info}
\alias{tp_get_info}
\title{Retrieve from the web basic information about the Telegram channel}
\usage{
tp_get_info(channel_id = NULL, channel_name = NULL)
}
\value{
A tibble.
}
\description{
Information retrieved includes channel title, description, number of
followers, and url to (cached) image.
}
