% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tp_get_metadata.R
\name{tp_get_metadata}
\alias{tp_get_metadata}
\title{Get (and cache) metadata for previously exported Telegram files}
\usage{
tp_get_metadata(path)
}
\arguments{
\item{path}{Path to a json file or a folder of json files Exported from Telegram. If a folder is given, all json files within it will be processed.}
}
\value{
A tibble with basic metadata about all json files.
}
\description{
Metadata are stored as csv files along with the json file.
}
