% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tp_set.R
\name{tp_get_options}
\alias{tp_get_options}
\title{Retrieve package options as set for the current session}
\usage{
tp_get_options(path = NULL)
}
\arguments{
\item{path}{Path to folder where Telegram exported archives are stored.}
}
\value{
A list with all options.
}
\description{
Retrieve package options as set for the current session
}
\examples{
tp_set_options(path = fs::path_home_r("telegramparser"))

tp_get_options()

tp_get_options(path = fs::path_home_r("other_tp"))
}
