% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tp_read_json.R
\name{tp_read_json}
\alias{tp_read_json}
\title{Reads a full Telegram file exported as json, and reads it in as a tibble}
\usage{
tp_read_json(path)
}
\arguments{
\item{path}{Path to the json file.}
}
\value{
A tibble with the full textual contents of the json files, including some columns as lists.
}
\description{
Reads a full Telegram file exported as json, and reads it in as a tibble
}
