% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tp_read_messages.R
\name{tp_read_messages}
\alias{tp_read_messages}
\title{Transform a full import of Telegram into a tibble with only main message textual contents}
\usage{
tp_read_messages(telegram_df)
}
\arguments{
\item{telegram_df}{A data frame, typically created with \code{\link[=tp_read_json]{tp_read_json()}}}
}
\value{
A tibble; none of the columns is a list.
}
\description{
Transform a full import of Telegram into a tibble with only main message textual contents
}
