% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tp_rename.R
\name{tp_rename}
\alias{tp_rename}
\title{Rename Telegram export files consistently}
\usage{
tp_rename(path = NULL, recurse = FALSE, preview = FALSE)
}
\arguments{
\item{path}{Path to folder where "result.json" files are stored.}

\item{recurse}{Defaults to FALSE. If TRUE, files in sub-folders are similarly
processed.}

\item{preview}{Defaults to FALSE. If TRUE}
}
\value{
If preview is set to TRUE, returns path of files to be renamed in a
tibble with two columns, \code{old_path} and \code{new_path}. If it is set to FALSE,
returns the same tibble, but invisibly.
}
\description{
Only json files with a file name starting with "result" will be renamed.
}
\examples{
\dontrun{
if (interactive) {
  tp_rename(path = ".", preview = "TRUE")
}
}
}
