% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tp_select.R
\name{tp_select}
\alias{tp_select}
\title{Select a Telegram channel from a folder where multiple archives are stored}
\usage{
tp_select(
  path = NULL,
  channel_name = NULL,
  channel_id = NULL,
  only_latest = TRUE
)
}
\arguments{
\item{path}{Path to a folder where Telegram archives in json format have
previously been stored.}

\item{channel_name}{Defaults to NULL. Name of one ore more channels, must
match exactly. Be aware that, in principle, different channels can have
exactly the same name or can change name, \code{channel_id} is generally safer.}

\item{channel_id}{Defaults to NULL. Numeric identifier of one ore more
channels, must match exactly.}
}
\value{
A vector of paths to json files corresponding to the relevant
Telegram channel.
}
\description{
Select a Telegram channel from a folder where multiple archives are stored
}
