% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tp_set_username.R
\name{tp_set_all_usernames}
\alias{tp_set_all_usernames}
\title{Interactively set all Telegram channel usernames not yet stored locally}
\usage{
tp_set_all_usernames(path = NULL)
}
\arguments{
\item{path}{Path to a folder where a \code{tp_usernames.csv} file is expected to
be located. This will usually be the same folder where Telegram archives
are stored.}
}
\description{
Interactively set all Telegram channel usernames not yet stored locally
}
\examples{
if (interactive()) {
  tp_set_all_usernames()
}
}
