% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tp_set.R
\name{tp_set_options}
\alias{tp_set_options}
\title{Set options for \code{telegramparser}}
\usage{
tp_set_options(path = NULL)
}
\arguments{
\item{path}{Path to folder where Telegram exported archives are stored.}
}
\value{
All options in a list.
}
\description{
Set options for \code{telegramparser}
}
\examples{
tp_set_options(fs::path_home_r(), "telegramparser")
}
