% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tp_set_username.R
\name{tp_set_username}
\alias{tp_set_username}
\title{Set the channel username for locally stored Telegram channels}
\usage{
tp_set_username(
  channel_username = NULL,
  channel_id = NULL,
  channel_name = NULL,
  path = NULL
)
}
\arguments{
\item{path}{Path to a folder where a \code{tp_usernames.csv} file is expected to
be located. This will usually be the same folder where Telegram archives
are stored.}
}
\value{
A data frame with one row and three columns (\code{channel_id},
\code{channel_name}, \code{channel_username}) if relevant data has previously been
stored. Otherwise, returns NULL.
}
\description{
Channel username is necessary to re-create links to channels and posts.
Unfortunately, it cannot be retrieved from the export json files themselves
and needs to be added manually.
}
\details{
Information is cached locally in a file called "tp_usernames.csv" located
in the path folder.

If no \code{channel_username} is given, the user will be requesed to add it
interactively.
}
\examples{
tp_set_username(
  channel_name = "BBC News | Русская служба",
  channel_username = "bbcrussian"
)
}
