% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tp_read_messages.R
\name{tp_read_messages}
\alias{tp_read_messages}
\title{Transform a full import of Telegram into a tibble with textual contents as plain character strings}
\usage{
tp_read_messages(telegram_df)
}
\arguments{
\item{telegram_df}{A data frame, typically created with \code{\link[=tp_read_json]{tp_read_json()}}}
}
\value{
A tibble; none of the columns is a list.
}
\description{
A unique \code{doc_id} column is added in order to provide a unique identifier when combinining different channels.
}
